requirements.txt
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
distro==1.9.0
dj-database-url==3.1.2
Django==6.0.3
docstring_parser==0.18.0
google-api-core==2.30.3
google-auth==2.49.1
google-cloud-aiplatform==1.149.0
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-cloud-resource-manager==1.17.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.68.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
packaging==26.2
pillow==12.1.1
proto-plus==1.27.2
protobuf==6.33.6
psycopg2==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
websockets==16.0



##.\venv\Scripts\activate
##py -m pip freeze > requirements.txt
##py -m pip list
##pip install -r requirements.txt
##py manage.py runserver